requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-anymail==10.3
django-cors-headers==3.14.0
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
environs==11.0.0
idna==3.7
import-export==0.3.1
inflection==0.5.1
marshmallow==3.21.3
packaging==24.0
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
stripe==9.10.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
